backend/requirements.txt
fastapi
uvicorn
websockets
pydantic-settings
google-genai
openai-whisper
python-multipart
edge-tts
